refactor: #46: move extractor functions to a one file
